Fix warnings for Search

+++ coding/trie.hpp
@@ -37,7 +37,7 @@ public:
   virtual ~Iterator() {}
 
   virtual Iterator<ValueT, EdgeValueT> * Clone() const = 0;
-  virtual Iterator<ValueT, EdgeValueT> * GoToEdge(uint32_t i) const = 0;
+  virtual Iterator<ValueT, EdgeValueT> * GoToEdge(size_t i) const = 0;
 };
 
 namespace reader
